feat: drag & drop now works correctly
