+++ src/constants/dnd.item.type.ts
@@ -1,0 +1,1 @@
+export const TASK_DND_TYPE = 'TASK'
